refactor(RichText): extract list-depth helpers to engine/listDepths
